Move callers to separate package
